% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.CFA.R
\name{coef.CFA}
\alias{coef.CFA}
\alias{coefficients}
\title{S3 coefficients for CFA}
\usage{
\method{coef}{CFA}(object, ...)
}
\arguments{
\item{object}{object of class \code{"CFA"}.}

\item{...}{other parameters passed trough.}
}
\value{
Coefficients extracted from the model object of class\code{"CFA"}.
}
\description{
S3 coefficients method for object of class\code{"CFA"}.
}
\keyword{methods}
