% Generated by roxygen2 (4.0.2): do not edit by hand
\name{S2CFA}
\alias{S2CFA}
\title{Configural Frequencies Analysis for two Samples.}
\usage{
S2CFA(patternfreq, alpha = 0.05, ccor = FALSE, ...)
}
\arguments{
\item{patternfreq}{an object of class \code{"Pfreq"}, which is data in pattern frequencies representation - see function \code{\link{dat2fre}}. The variable defining the two subsaples (a variable with max. two categories) must be located in the last but one column of the object of class \code{"Pfreq"}}

\item{alpha}{a numeric giving the alpha level for testing (default set to \code{alpha=.05})}

\item{ccor}{a logical (TRUE / FALSE) determining wether to apply a continuity correction or not. When set to \code{ccor=TRUE} continuity correction is applied. For \code{ccor=FALSE} no continuity correction is applied.}

\item{...}{additional parameters passed through to other functions.}
}
\value{
an object of class \code{S2CFA} with results.
}
\description{
Calculates coefficients for the two-sample CFA. Instead of differentiating between types and antitypes, two-sample CFA looks for discrimination types, that is configurations with significant differences in frequencies between two subsamples.
}
\details{
no details at the moment ...
}
\examples{
#######################################
############### some examples #########
######### example from Marks Textbook
data(Lienert1978)
res1 <- S2CFA(Lienert1978)
summary(res1)
res2 <- S2CFA(Lienert1978, ccor=TRUE) # with continuity correction
summary(res2)
######### example with biger numbers
data(suicide)
ftab(suicide) # 'Epoche' may divide the sample into 2 subsamples
suicide_2s <- suicide[, c(1,3,2) ] # reorder data that 'Epoche' is the last column
ftab(suicide_2s) # check reordering
suicide_2s_fre <- dat2fre(suicide_2s)
res3 <- S2CFA(suicide_2s_fre)
summary(res3)
res4 <- S2CFA(suicide_2s_fre, ccor=TRUE) # with continuity correction
summary(res4)
}
\references{
Stemmler, M. (2014). \emph{Person-Centered Methods – Configural Frequency Analysis (CFA) and Other Methods for the Analysis of Contingency Tables.} Cham Heidelberg New York Dordrecht London: Springer.

Stemmler, M., & Hammond, S. (1997). Configural frequency analysis of dependent samples for intra-patient treatment comparisons. \emph{Studia Psychologica, 39}, 167–175.
}

