% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.CFA}
\alias{summary.CFA}
\title{S3 Summary for CFA}
\usage{
\method{summary}{CFA}(object, digits = 3, type = "z.pChi", ...)
}
\arguments{
\item{object}{object of class\code{"CFA"}}

\item{digits}{integer rounds the values to the specified number of decimal places, default is \code{digits=3}.}

\item{type}{character with default \code{type="z.pChi"}, to return wether the observed pattern are 'Types', 'Antitypes' or not significant at all. Possible options for \code{type} are \code{"pChi"}, \code{"ex.bin.test"}, \code{"z.pChi"}, \code{"z.pBin"} and \code{"p.stir"}.}

\item{...}{other parameters passed trough}
}
\description{
S3 summary method for object of class\code{"CFA"}
}

