% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.S2CFA.R
\name{summary.S2CFA}
\alias{summary.S2CFA}
\title{S3 Summary for S2CFA}
\usage{
\method{summary}{S2CFA}(object, digits = 3, type = "ex.fisher.test",
  ...)
}
\arguments{
\item{object}{object of class\code{"S2CFA"}}

\item{digits}{integer rounds the values to the specified number of decimal places, default is \code{digits=3}.}

\item{type}{character with default \code{type="ex.fisher.test"}, to return wether the observed pattern are 'discriminating Types' or not significant at all based on the respective p-value. Another option for \code{type} is \code{type="pChi"}.}

\item{...}{other parameters passed trough}
}
\description{
S3 summary method for object of class\code{"S2CFA"}
}
